Verbatim checkin of example code from https://github.com/hypfvieh/bluez-dbus/issues/1#issuecomment-389770063
As of this checkin the file fails to build with two errors.
The first error is that the package directive at the start of the file
does not match the directory it has been added to,  This will be trivial to fix.
The second error appears likely to reflect a change in the API provided by the DBusConnection class
since the example code was written in 2018.  Hopefully this can also be addressed