Commenting out explicit version dependencies
These were added at an earlier time to silence GH dependabot security
warnings.
They needed to be in the top level build.gradle because this is where
com.android.tools.build:gradle is added (i.e. the Android Gradle Plugin),
and the insecure versions were brought in by the version of that package
at that time.

Hopefully, with version 8.13.2 of AGP, all insecure dependencies have
(for the time being) been resolved, and it is OK to remove the
explicit dependencies on fixed versions.